+++ src/ios/WKWebViewFix/CDVWKWebViewEngine+HCPPlugin_ReadAccessURL.h
@@ -10,20 +10,14 @@
     #import "CDVWKWebViewEngine.h"
     #define FOUND_WK_WEB_VIEW 1
     @interface CDVWKWebViewEngine (HCPPlugin_ReadAccessURL)
+    @end
 #else
     #if CORDOVA_VERSION_MIN_REQUIRED >= __CORDOVA_6_1_0
         #define FOUND_WK_WEB_VIEW 1
         #import "../../../../CordovaLib/Classes/Private/Plugins/CDVWebViewEngine/CDVWebViewEngine.h"
         @interface CDVWebViewEngine (HCPPlugin_ReadAccessURL)
+        @end
     #else
         #warning CANNOT FIND ANY WebViewEngine Include
     #endif
-#endif
-
-#if FOUND_WK_WEB_VIEW==1
-
-- (id)loadRequest:(NSURLRequest*)request;
-
-@end
-
 #endif